--- crypto/math/src/fft/cpu/mod.rs
@@ -1,8 +1,0 @@
-pub mod bit_reversing;
-#[cfg(feature = "alloc")]
-pub mod bowers_fft;
-#[cfg(all(test, feature = "alloc"))]
-mod bowers_fft_tests;
-pub mod fft;
-#[cfg(feature = "alloc")]
-pub mod roots_of_unity;
