+++ crypto/math/src/fft/mod.rs
@@ -1,7 +1,15 @@
-pub mod cpu;
+pub mod bit_reversing;
+#[cfg(feature = "alloc")]
+pub mod bowers_fft;
 pub mod errors;
+pub mod fft;
 #[cfg(feature = "alloc")]
 pub mod polynomial;
+#[cfg(feature = "alloc")]
+pub mod roots_of_unity;
+
+#[cfg(all(test, feature = "alloc"))]
+mod bowers_fft_tests;
 
 #[cfg(all(test, feature = "alloc"))]
 pub(crate) mod test_helpers;
