+++ crypto/math/src/fft/test_helpers.rs
@@ -1,5 +1,5 @@
 use crate::{
-    fft::cpu::roots_of_unity::get_powers_of_primitive_root,
+    fft::roots_of_unity::get_powers_of_primitive_root,
     field::{
         element::FieldElement,
         traits::{IsFFTField, RootsConfig},
