+++ crypto/stark/src/domain.rs
@@ -1,5 +1,5 @@
 use math::{
-    fft::cpu::roots_of_unity::get_powers_of_primitive_root_coset,
+    fft::roots_of_unity::get_powers_of_primitive_root_coset,
     field::{
         element::FieldElement,
         traits::{IsFFTField, IsField, IsSubFieldOf},
