+++ crypto/stark/src/fri/fri_functions.rs
@@ -1,4 +1,4 @@
-use math::fft::cpu::{
+use math::fft::{
     bit_reversing::in_place_bit_reverse_permute, roots_of_unity::get_powers_of_primitive_root_coset,
 };
 use math::field::{
